Elastic APM module, with .ejs for lodash template
