+++ modules/elastic-apm/plugin.js
@@ -1,0 +1,7 @@
+import { init as initApm } from '@elastic/apm-rum';
+
+import options from './options';
+
+export default () => {
+  initApm(options);
+};
