+++ modules/elastic-apm/templates/options.ejs
@@ -1,0 +1,1 @@
+export default <%= JSON.stringify(options, null, 2) %>;
